% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.power.longtest.R
\name{power.longtest}
\alias{power.longtest}
\title{Constructor function for class \code{"power.longtest"}}
\usage{
power.longtest(object)
}
\arguments{
\item{object}{a list.}
}
\value{
an object of class \code{"power.longtest"}
}
\description{
Constructor function for class \code{"power.longtest"}
}
