% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcrm.power.R
\name{rcrm.power}
\alias{rcrm.power}
\title{Power for Random Coefficient Regression Model (RCRM)}
\usage{
rcrm.power(
  n = NULL,
  delta = NULL,
  power = NULL,
  t = NULL,
  lambda = 1,
  sig2.i = 0,
  cor.s.i = NULL,
  sig2.s = 0,
  sig2.e = NULL,
  sig.level = 0.05,
  droprate = NULL,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^2
)
}
\arguments{
\item{n}{sample size, group 1}

\item{delta}{Effect size (absolute difference in rate of decline between tx and placebo)}

\item{power}{power}

\item{t}{Vector of visit time points (including time 0)}

\item{lambda}{allocation ratio (= (sample size group 1)/(sample size group 2)}

\item{sig2.i}{Variance of random intercept}

\item{cor.s.i}{Correlation between random intercept & slope}

\item{sig2.s}{Variance of random slope}

\item{sig2.e}{Variance of pure error}

\item{sig.level}{type one error}

\item{droprate}{Exponential dropout rate}

\item{alternative}{one- or two-sided test}

\item{tol}{numerical tolerance used in root finding}
}
\value{
One of the number of subject required per arm, the `power`, or
detectable effect size given `sig.level` and the other parameter estimates.
}
\description{
The RCRM theoretical results are summarized in a manuscript (Hu, Mackey, and
Thomas; 2021).
}
\details{
This program was co-developed by Nan Hu (Genentech Inc., Biostatistics)
and a summer intern Zhe Qu (Ph.D. candidate at Tulane University). If you
have any technical questions or any ideas for improving functionality,
please contact Nan Hu (hu.nan@gene.com).
}
\examples{
\dontrun{
browseVignettes(package = "longpower")
}
# An Alzheimer's Disease example using ADAS-cog pilot estimates
t <- seq(0,1.5,0.25)

rcrm.power(delta=1.5, t=t, sig2.s = 24, sig2.e = 10, cor.s.i=0.5, 
  sig.level=0.05, power = 0.80)
rcrm.power(n=180, t=t, sig2.s = 24, sig2.e = 10, cor.s.i=0.5, 
  sig.level=0.05, power = 0.80)
rcrm.power(n=180, delta=1.5, t=t, sig2.s = 24, sig2.e = 10, 
  cor.s.i=0.5, sig.level=0.05)

rcrm.power(delta=1.5, t=t, sig2.s = 24, sig2.e = 10, cor.s.i=0.5, 
  sig.level=0.05, power = 0.80, alternative = 'one.sided')
rcrm.power(n=142, t=t, sig2.s = 24, sig2.e = 10, cor.s.i=0.5, 
  sig.level=0.05, power = 0.80, alternative = 'one.sided')
rcrm.power(n=142, delta=1.5, t=t, sig2.s = 24, sig2.e = 10, 
  cor.s.i=0.5, sig.level=0.05, alternative = 'one.sided')

}
\references{
Hu, N., Mackey, H., & Thomas, R. (2021). Power and sample size
for random coefficient regression models in randomized experiments with
monotone missing data. \emph{Biometrical Journal}, 63(4), 806-824.
}
\seealso{
\code{\link{lmmpower}}, \code{\link{edland.linear.power}}, 
\code{\link{two.stage.me.power}}
}
\author{
Nan Hu, Michael C. Donohue
}
