% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two.stage.me.power.R
\name{two.stage.me.power}
\alias{two.stage.me.power}
\title{Power for two-stage mixed effects model}
\usage{
two.stage.me.power(
  n = NULL,
  delta = NULL,
  power = NULL,
  duration = NULL,
  n_assessments = NULL,
  lambda = 1,
  sig2.b = 1,
  sig2.w = 1,
  sig.level = 0.05,
  alternative = c("two.sided", "one.sided"),
  tol = .Machine$double.eps^2
)
}
\arguments{
\item{n}{sample size, group 1}

\item{delta}{Effect size (absolute difference in rate of decline between tx and placebo)}

\item{power}{power}

\item{duration}{Duration of follow-up}

\item{n_assessments}{Number of assessments (including baseline)}

\item{lambda}{allocation ratio (= (sample size group 1)/(sample size group 2)}

\item{sig2.b}{Between-subject variance}

\item{sig2.w}{Within-subject variance}

\item{sig.level}{type one error}

\item{alternative}{one- or two-sided test}

\item{tol}{numerical tolerance used in root finding}
}
\value{
One of the number of subject required per arm, the `power`, or
detectable effect size given `sig.level` and the other parameter estimates.
}
\description{
The two-stage mixed effects model details can be found
in Section 8.4 of Fitzmaurice, Laird, and Ware (2011).
}
\details{
This program was co-developed by Nan Hu (Genentech Inc., Biostatistics)
and a summer intern Zhe Qu (Ph.D. candidate at Tulane University). If you
have any technical questions or any ideas for improving functionality,
please contact Nan Hu (hu.nan@gene.com).
}
\examples{
\dontrun{
browseVignettes(package = "longpower")
}
# An Alzheimer's Disease example using ADAS-cog pilot estimates

two.stage.me.power(delta=1.5, duration = 1.5, n_assessments = 7, sig2.b = 24, 
  sig2.w = 10, sig.level=0.05, power = 0.80)
two.stage.me.power(n=207, duration = 1.5, n_assessments = 7, sig2.b = 24, 
  sig2.w = 10, sig.level=0.05, power = 0.80)
two.stage.me.power(n=207, delta=1.5, duration = 1.5, n_assessments = 7, 
  sig2.b = 24, sig2.w = 10, sig.level=0.05)

}
\references{
Hu, N., Mackey, H., & Thomas, R. (2021). Power and sample size
for random coefficient regression models in randomized experiments with
monotone missing data. \emph{Biometrical Journal}, 63(4), 806-824.
}
\seealso{
\code{\link{lmmpower}}, \code{\link{edland.linear.power}}, 
\code{\link{rcrm.power}}
}
\author{
Nan Hu, Michael C. Donohue
}
